Sanitize Exceptions.
Exception was broadly used for no reason and CoreException was catched
and ignored in a number of places. This commit simply adds throws for
the exact exceptions.
Hopefully this will allow us to find some test failures that appear
randomly based on order of running (probably not cleaned up resources.
All in all Exception should not be catched/called anywhere as it has to
nasty side effects.