Never deal with Exception directly- only with subclasses.
Catching Exception in production code leads to nasty effects sometimes
as it also catches RuntimeException which have an unpredictable side
effects.
Also catching Exception and failing in a unit test is bad because:
1. junit will properly mark the test as failing even if it's not
catched
2. all the reasons why it failed are lost as it is threated as smth
expected and if you catch the exception it's your task to print the
stack trace and etc.