Simplify the abstract handler.
There is no point in getting the projectRoot as the very first thing in
the Handler.execute, set it to local variable in the parent and request
it via getter after that in the very same execute method. Stick to local
variable instead as it's less errorpron because someone could have
called getProjecRoot but forgot to set it before that.